# processes/src/main/resources/schema/processes/openapi.yaml
openapi: 3.0.3
info:
  title: "A sample API conforming to version 1.0.0 of the OGC API - Processes - Part 1: Core Standard"
  description: "A sample API conforming to version 1.0.0 of the [OGC API - Processes - Part 1: Core](https://docs.ogc.org/is/18-062r2/18-062r2.html) Standard.\n\n
    \ Copyright (c) 2021 Open Geospatial Consortium.\n\n
    \ To obtain additional rights of use, visit http://www.opengeospatial.org/legal/ \n\n
    \ This document is an API definition document provided alongside the
    \ OGC API - Processes standard. The OGC API - Processes Standard specifies a processing
    \ interface to communicate over a RESTful protocol using JavaScript Object Notation
    \ (JSON) encodings. The specification allows for the wrapping of computational tasks
    \ into executable processes that can be offered by a server and be invoked by a
    \ client application."
  contact:
    name: Open Geospatial Consortium (OGC)
    url: https://www.ogc.org/contacts
    email: standards-team@ogc.org
  license:
    name: OGC license
    url: http://www.ogc.org/legal/
  version: 1.0.0
servers:
  - url: http://example.org/ogcapi
    description: Example server
paths:
...
  /processes/{processID}:
    get:
      tags:
        - ProcessDescription
      summary: retrieve a process description
      description: "The process description contains information about inputs and\
        \ outputs and a link to the execution-endpoint for the process. The Core does\
        \ not mandate the use of a specific process description to specify the interface\
        \ of a process. That said, the Core requirements class makes the following\
        \ recommendation:\n\nImplementations SHOULD consider supporting the OGC process\
        \ description.\n\nFor more information, see [Section 7.10](https://docs.ogc.org/is/18-062/18-062.html#sc_process_description).\n"
      operationId: getProcessDescription
      parameters:
        - name: processID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A process description.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/process'
              examples:
                processDecription:
                  summary: A Process Description example
                  externalValue: ../../examples/json/ProcessDescription.json
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
  /jobs:
    get:
      tags:
        - JobList
      summary: retrieve the list of jobs.
      description: "Lists available jobs.\n\nFor more information, see [Section 11](https://docs.ogc.org/is/18-062/18-062.html#sc_job_list).\n"
      operationId: getJobs
      responses:
        "200":
          description: A list of jobs for this process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobList'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "501":
          description: Listing jobs is not implemented.
  /processes/{processID}/execution:
    post:
      tags:
        - Execute
      summary: execute a process.
      description: "Create a new job.\n\nFor more information, see [Section 7.11](https://docs.ogc.org/is/18-062/18-062.html#sc_create_job).\n"
      operationId: execute
      parameters:
        - name: processID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Mandatory execute request JSON
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/execute'
        required: true
      responses:
        "200":
          description: Result of synchronous execution
          content:
            /*:
              schema:
                $ref: '#/components/schemas/inline_response_200'
              example:
                message:
                  message: "Job submitted with ID: 123e4567-e89b-12d3-a456-426614174000"
                status:
                  message: "200"
                jobID: 123e4567-e89b-12d3-a456-426614174000
        "201":
          description: Started asynchronous execution. Created job.
          headers:
            Location:
              description: URL to check the status of the execution/job.
              schema:
                type: string
            Preference-Applied:
              description: The preference applied to execute the process asynchronously
                (see. RFC 2740).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "500":
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
      callbacks:
        jobCompleted:
          '{$request.body#/subscriber/successUri}':
            post:
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/results'
              responses:
                "200":
                  description: Results received successfully
  /jobs/{jobId}:
    get:
      tags:
        - Status
      summary: retrieve the status of a job
      description: "Shows the status of a job.\n\n For more information, see [Section\
        \ 7.12](https://docs.ogc.org/is/18-062/18-062.html#sc_retrieve_status_info).\n"
      operationId: getStatus
      parameters:
        - name: jobId
          in: path
          description: local identifier of a job
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The status of a job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "500":
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
    delete:
      tags:
        - Dismiss
      summary: "cancel a job execution, remove a finished job"
      description: "Cancel a job execution and remove it from the jobs list.\n\nFor\
        \ more information, see [Section 13](https://docs.ogc.org/is/18-062/18-062.html#Dismiss).\n"
      operationId: dismiss
      parameters:
        - name: jobId
          in: path
          description: local identifier of a job
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The status of a job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/statusInfo'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "500":
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "501":
          description: Dismissing jobs is not implemented.
  /jobs/{jobId}/results:
    get:
      tags:
        - Result
      summary: retrieve the result(s) of a job
      description: "Lists available results of a job. In case of a failure, lists\
        \ exceptions instead.\n\nFor more information, see [Section 7.13](https://docs.ogc.org/is/18-062/18-062.html#sc_retrieve_job_results).\n"
      operationId: getResult
      parameters:
        - name: jobId
          in: path
          description: local identifier of a job
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The results of a job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/results'
        "404":
          description: The requested URI was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "500":
          description: A server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
        "501":
          description: Retrieving job results is not implemented.
components:
  schemas:
...
    exception:
      title: Exception Schema
      required:
        - type
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      additionalProperties: true
      description: JSON schema for exceptions based on RFC 7807
...
    process:
      allOf:
        - $ref: '#/components/schemas/processSummary'
        - type: object
          properties:
            inputs:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/inputDescription'
            outputs:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/outputDescription'
    jobList:
      required:
        - jobs
        - links
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/statusInfo'
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    execute:
      type: object
      properties:
        inputs:
          type: object
          additionalProperties:
            oneOf:
              - $ref: '#/components/schemas/inlineOrRefData'
              - type: array
                items:
                  $ref: '#/components/schemas/inlineOrRefData'
        outputs:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/output'
        response:
          type: string
          default: raw
          enum:
            - raw
            - document
        subscriber:
          $ref: '#/components/schemas/subscriber'
    results:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/inlineOrRefData'
    statusInfo:
      required:
        - jobID
        - status
        - type
      type: object
      properties:
        processID:
          type: string
        type:
          type: string
          enum:
            - process
        jobID:
          type: string
        status:
          $ref: '#/components/schemas/statusCode'
        message:
          type: string
        collection:
          type: string
          description: Display name of the requested collection.
        dataSelection:
          type: string
          description: Dataset key or data selection requested for the download.
        format:
          type: string
          description: Requested output document format.
        metadataUrl:
          type: string
          description: Link to the metadata page supplied when the download was submitted.
        created:
          type: string
          format: date-time
        started:
          type: string
          format: date-time
        finished:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        progress:
          maximum: 100
          minimum: 0
          type: integer
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
    link:
      required:
        - href
      type: object
      properties:
        href:
          type: string
        rel:
          type: string
          example: service
        type:
          type: string
          example: application/json
        hreflang:
          type: string
          example: en
        title:
          type: string
    processSummary:
      allOf:
        - $ref: '#/components/schemas/descriptionType'
        - required:
            - id
            - version
          type: object
          properties:
            id:
              type: string
            version:
              type: string
            jobControlOptions:
              type: array
              items:
                $ref: '#/components/schemas/jobControlOptions'
            outputTransmission:
              type: array
              items:
                $ref: '#/components/schemas/transmissionMode'
            links:
              type: array
              items:
                $ref: '#/components/schemas/link'
    inputDescription:
      allOf:
        - $ref: '#/components/schemas/descriptionType'
        - required:
            - schema
          type: object
          properties:
            minOccurs:
              type: integer
              default: 1
            maxOccurs:
              oneOf:
                - type: integer
                  default: 1
                - type: string
                  enum:
                    - unbounded
            schema:
              $ref: '#/components/schemas/schema'
    outputDescription:
      allOf:
        - $ref: '#/components/schemas/descriptionType'
        - required:
            - schema
          type: object
          properties:
            schema:
              $ref: '#/components/schemas/schema'
    inlineOrRefData:
      oneOf:
        - $ref: '#/components/schemas/inputValueNoObject'
        - $ref: '#/components/schemas/qualifiedInputValue'
        - $ref: '#/components/schemas/link'
    output:
      type: object
      properties:
        format:
          $ref: '#/components/schemas/format'
        transmissionMode:
          $ref: '#/components/schemas/transmissionMode'
    subscriber:
      required:
        - successUrl
      type: object
      properties:
        successUri:
          type: string
          format: uri
        inProgressUri:
          type: string
          format: uri
        failedUri:
          type: string
          format: uri
      description: "Optional URIs for callbacks for this job.\n\nSupport for this\
        \ parameter is not required and the parameter may be\nremoved from the API\
        \ definition, if conformance class **'callback'**\nis not listed in the conformance\
        \ declaration under `/conformance`."
    statusCode:
      type: string
      nullable: false
      enum:
        - accepted
        - running
        - successful
        - failed
        - dismissed
    descriptionType:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/metadata'
        additionalParameters:
          allOf:
            - $ref: '#/components/schemas/metadata'
            - type: object
              properties:
                parameters:
                  type: array
                  items:
                    $ref: '#/components/schemas/additionalParameter'
    jobControlOptions:
      type: string
      enum:
        - sync-execute
        - async-execute
        - dismiss
    transmissionMode:
      type: string
      default: value
      enum:
        - value
        - reference
    schema:
      oneOf:
        - $ref: '#/components/schemas/reference'
        - type: object
          properties:
            title:
              type: string
            multipleOf:
              minimum: 0
              exclusiveMinimum: true
              type: number
            maximum:
              type: number
            exclusiveMaximum:
              type: boolean
              default: false
            minimum:
              type: number
            exclusiveMinimum:
              type: boolean
              default: false
            maxLength:
              minimum: 0
              type: integer
            minLength:
              minimum: 0
              type: integer
              default: 0
            pattern:
              type: string
              format: regex
            maxItems:
              minimum: 0
              type: integer
            minItems:
              minimum: 0
              type: integer
              default: 0
            uniqueItems:
              type: boolean
              default: false
            maxProperties:
              minimum: 0
              type: integer
            minProperties:
              minimum: 0
              type: integer
              default: 0
            required:
              minItems: 1
              uniqueItems: true
              type: array
              items:
                type: string
            enum:
              minItems: 1
              uniqueItems: false
              type: array
              items:
                type: object
            type:
              type: string
              enum:
                - array
                - boolean
                - integer
                - number
                - object
                - string
            not:
              oneOf:
                - $ref: '#/components/schemas/schema'
                - $ref: '#/components/schemas/reference'
            allOf:
              type: array
              items:
                oneOf:
                  - $ref: '#/components/schemas/schema'
                  - $ref: '#/components/schemas/reference'
            oneOf:
              type: array
              items:
                oneOf:
                  - $ref: '#/components/schemas/schema'
                  - $ref: '#/components/schemas/reference'
            anyOf:
              type: array
              items:
                oneOf:
                  - $ref: '#/components/schemas/schema'
                  - $ref: '#/components/schemas/reference'
            items:
              oneOf:
                - $ref: '#/components/schemas/schema'
                - $ref: '#/components/schemas/reference'
            properties:
              type: object
              additionalProperties:
                oneOf:
                  - $ref: '#/components/schemas/schema'
                  - $ref: '#/components/schemas/reference'
            additionalProperties:
              oneOf:
                - $ref: '#/components/schemas/schema'
                - $ref: '#/components/schemas/reference'
                - type: boolean
              default: true
            description:
              type: string
            format:
              type: string
            default:
              type: object
            nullable:
              type: boolean
              default: false
            readOnly:
              type: boolean
              default: false
            writeOnly:
              type: boolean
              default: false
            example:
              type: object
            deprecated:
              type: boolean
              default: false
            contentMediaType:
              type: string
            contentEncoding:
              type: string
            contentSchema:
              type: string
          additionalProperties: false
    inputValueNoObject:
      oneOf:
        - type: string
        - type: number
        - type: integer
        - type: boolean
        - type: array
          items:
            type: string
        # Change and embed directly as $ref byte not works with generator
        - type: string
          format: byte
        - $ref: '#/components/schemas/bbox'
    qualifiedInputValue:
      allOf:
        - $ref: '#/components/schemas/format'
        - required:
            - value
          type: object
          properties:
            value:
              $ref: '#/components/schemas/inputValue'
    format:
      type: object
      properties:
        mediaType:
          type: string
        encoding:
          type: string
        schema:
          oneOf:
            - type: string
              format: url
            - type: object
    metadata:
      type: object
      properties:
        title:
          type: string
        role:
          type: string
        href:
          type: string
    additionalParameter:
      required:
        - name
        - value
      type: object
      properties:
        name:
          type: string
        value:
          type: array
          items:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: array
                items:
                  type: object
              - type: object
    reference:
      required:
        - $ref
      type: object
      properties:
        $ref:
          type: string
          format: uri-reference
...
    bbox:
      required:
        - bbox
      type: object
      properties:
        bbox:
          type: array
          oneOf:
            - maxItems: 4
              minItems: 4
              type: object
            - maxItems: 6
              minItems: 6
              type: object
          items:
            type: number
        crs:
          type: string
          format: uri
          default: http://www.opengis.net/def/crs/OGC/1.3/CRS84
          enum:
            - http://www.opengis.net/def/crs/OGC/1.3/CRS84
            - http://www.opengis.net/def/crs/OGC/0/CRS84h
    inputValue:
      oneOf:
        - $ref: '#/components/schemas/inputValueNoObject'
        - type: object
    inline_response_200:
      oneOf:
        - type: string
        - type: number
        - type: integer
        - type: object
          nullable: true
        - type: array
          items:
            type: object
        - type: boolean
        - type: string
          format: binary
        - $ref: '#/components/schemas/results'
...